+++ pkg/promotel/internal/exporter_test.go
@@ -1,4 +1,4 @@
-package promotel
+package internal
 
 import (
 	"context"
