+++ pkg/promotel/internal/receiver.go
@@ -1,4 +1,4 @@
-package promotel
+package internal
 
 import (
 	"context"
@@ -11,7 +11,6 @@ import (
 	"go.opentelemetry.io/collector/receiver"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
-	"github.com/smartcontractkit/chainlink-common/pkg/promotel/internal"
 )
 
 type Runnable interface {
@@ -38,7 +37,7 @@ func (p *metricReceiver) Close() error {
 	return p.receiver.Shutdown(context.Background())
 }
 
-func NewMetricReceiver(config ReceiverConfig, g prometheus.Gatherer, r prometheus.Registerer, logger logger.Logger, next NextFunc) (Runnable, error) {
+func NewMetricReceiver(config *ReceiverConfig, g prometheus.Gatherer, r prometheus.Registerer, logger logger.Logger, next NextFunc) (Runnable, error) {
 	factory := prometheusreceiver.NewFactoryWithOptions(
 		prometheusreceiver.WithGatherer(g),
 		prometheusreceiver.WithRegisterer(r),
@@ -46,19 +45,19 @@ func NewMetricReceiver(config ReceiverConfig, g prometheus.Gatherer, r prometheu
 	// Creates a metrics receiver with the context, settings, config, and consumer
 	receiver, err := factory.CreateMetrics(
 		context.Background(),
-		internal.NewReceiverSettings(logger),
+		NewReceiverSettings(logger),
 		config,
-		internal.NewConsumer(consumer.ConsumeMetricsFunc(next)))
+		NewConsumer(consumer.ConsumeMetricsFunc(next)))
 	if err != nil {
 		return nil, err
 	}
 	// Creates a no-operation host for the receiver
-	host := internal.NewNopHost()
+	host := NewNopHost()
 	return &metricReceiver{factory, host, receiver, logger}, nil
 }
 
-func NewDebugMetricReceiver(config ReceiverConfig, g prometheus.Gatherer, r prometheus.Registerer, logger logger.Logger) (MetricReceiver, error) {
-	debugExporter := internal.NewDebugExporter(logger)
+func NewDebugMetricReceiver(config *ReceiverConfig, g prometheus.Gatherer, r prometheus.Registerer, logger logger.Logger) (MetricReceiver, error) {
+	debugExporter := NewDebugExporter(logger)
 	// Creates a no-operation consumer
 	return NewMetricReceiver(config, g, r, logger, func(_ context.Context, md pmetric.Metrics) error {
 		// Writes metrics data to stdout
